backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
]
